requirements.txt
streamlit==1.24.0
nltk==3.8.1
pandas
matplotlib
numpy
scikit-learn
plotly
langchain
openai==0.28
setuptools
langchain
python-dotenv
pandasai
chromadb